requirements.txt
requests==2.25.1
pandas==1.2.0
yagmail==0.14.245
python-dotenv==0.19.2
xlrd2==1.3.4